+++ HISTORY.rst
@@ -211,6 +211,8 @@
   template.
 - Remove converter.validators and replace it with common.validators.
 - Autoimport search.py modules from apps.
+- Make ``SearchField`` label optional. If not specified, the ``verbose_name``
+  of the model field will be used instead.
 
 3.4.14 (2020-08-18)
 ===================